+++ CHANGELOG.md
@@ -61,6 +61,7 @@ v5.3.0
 - Replaced depricated "sprintf" with "std::to_string" (!1406)
 - Extended Reynolds Stresses filter to passive scalars (!1430)
 - Fixed Taylor-Hood expansion for VCSWeakPressure (!1444)
+- Fix filename in LinearisedAdvection (!1479)
 
 **VortexWaveInteractionSolver**
 - Replaced depricated "sprintf" with "std::to_string" (!1406)